requirements.txt
numpy==1.24.3
pandas==2.0.2
yfinance==0.2.31
python-binance==1.0.19
ta-lib==0.4.28
scikit-learn==1.2.2
pandas-ta==0.3.14b0
python-dotenv==1.0.0
websocket-client==1.6.1
aiohttp==3.8.5
asyncio==3.4.3
ccxt==4.0.0
fastapi==0.100.0
uvicorn==0.22.0
redis==4.6.0
pytest==7.4.0
logging==0.4.9.6
